Ensure subtest results are reported no matter what
Subtests with TIMEOUT or NOTRUN results could cause the subtest not to be
reported at all, even though one report had a PASS for that subtest. This
mechanism had been initially included to avoid *false negatives*, but now that
the test runner much more stable, it actually has the opposite effect and
actually fails to report *true negatives*. In particular, it seems important to
report subtests that time out across all tested implementations in the
"complete-fails.html" report, so that they do not fall through the cracks.

The consolidation process also created a fake subtest when no subtests were
found for a given test. This is useful for most manual tests and for reftests
where the top level test *is* the subtest. However, in some cases, a report may
contain a FAIL or TIMEOUT for a test without reporting on individual subtests,
perhaps because the report has not been generated with the latest version of the
tests, or perhaps because the test times out before being able to report on
subtests. When that happens, it seems preferable to report the test result on
the test line and report on the actual subtests (with a "-" outcome for the
agent that does not contain any subtest results) without creating a fake subtest
line. This commit does precisely that: it builds an index of tests that contain
subtests and uses that information during the consolidation process to determine
whether a fake subtest is needed.

Note that the fake subtest line was often (although not always!) filtered out
by the "false negatives" mechanism described above.