Added processing for message comments
If a message comment on a test or a subtest contains the string
"NOTRUN:" or "SKIPPED:" then the result of the test or subtest is
not used nor reported in the overall calculations. Useful for subtests
that are not run on some platforms (e.g., when an A11Y Technology API
does not support some minor aspect of an overall feature.

Note that the result of a test case will never be PASS if all of the
subtests are skipped.  In that case the result of the test will be
NOTRUN.